refactor: move pre v7 fw tests to seperate test file.
